Feature/docker files (#1)
* Refactor into one script.

* Update docker-compose and readme.
