+++ Dockerfile
@@ -2,10 +2,10 @@ FROM python:3.7-alpine
 
 WORKDIR /usr/src/app
 
+ENV MODE=mq_produce
 COPY requirements.txt ./
 RUN pip install --no-cache-dir -r requirements.txt
 
 COPY . .
 
-CMD [ "python", "./mq_producer.py" ]
-
+CMD python ./rbmq-test.py ${MODE}
